refactored locker container
