+++ src/main/scala/net/psforever/objects/avatar/Avatar.scala
@@ -3,7 +3,8 @@ package net.psforever.objects.avatar
 import net.psforever.objects.definition.{AvatarDefinition, BasicDefinition}
 import net.psforever.objects.equipment.{EquipmentSize, EquipmentSlot}
 import net.psforever.objects.loadouts.{Loadout, SquadLoadout}
-import net.psforever.objects.{GlobalDefinitions, LockerContainer, LockerEquipment, OffhandEquipmentSlot}
+import net.psforever.objects.locker.{LockerContainer, LockerEquipment}
+import net.psforever.objects.{GlobalDefinitions, OffhandEquipmentSlot}
 import net.psforever.types._
 import org.joda.time.{Duration, LocalDateTime, Seconds}
 
@@ -82,7 +83,7 @@ case class Avatar(
     loadouts: Seq[Option[Loadout]] = Seq.fill(15)(None),
     squadLoadouts: Seq[Option[SquadLoadout]] = Seq.fill(10)(None),
     implants: Seq[Option[Implant]] = Seq(None, None, None),
-    locker: LockerContainer = new LockerContainer(),          // TODO var bad
+    locker: LockerContainer = LockerContainer(),          // TODO var bad
     deployables: DeployableToolbox = new DeployableToolbox(), // TODO var bad
     lookingForSquad: Boolean = false,
     var vehicle: Option[PlanetSideGUID] = None, // TODO var bad
