+++ src/main/scala/net/psforever/objects/inventory/Container.scala
@@ -98,94 +98,3 @@ trait Container {
   def Collisions(index: Int, width: Int, height: Int): Try[List[InventoryItem]] =
     Inventory.CheckCollisionsVar(index, width, height)
 }
-
-//object Container {
-//  type ValidContainer = PlanetSideServerObject with Container
-//
-//  final case class GetMoveItem(where_src : Int, other : ValidContainer, where_other : Int, other_item : Option[Equipment])
-//
-//  final case class GiveMoveItem(cont1 : ValidContainer, where_cont1 : Int, item : Option[Equipment], cont2 : ValidContainer, where_cont2 : Int, other_item : Option[Equipment])
-//
-//
-//  final case class TakeMoveItem(source_index : Int, destination : ValidContainer, destination_index : Int)
-//
-//  final case class TakeMoveItem2(item_guid : PlanetSideGUID, destination : ValidContainer, destination_index : Int)
-//
-//  final case class GivingMoveItem(item : Equipment, source : ValidContainer, source_index : Int, destination : ValidContainer, destination_index : Int)
-//
-//  final case class PutMoveItem(item : Equipment, target_index : Int, source : ValidContainer, source_index : Int)
-//
-//  final case class DropMoveItem(item : Equipment, source : ValidContainer, source_index : Int)
-//
-//  final case class ItemMoved(item : Equipment, location : ValidContainer, location_index : Int)
-//
-//  final case class NoMoveItem(source : ValidContainer, source_index : Int)
-//}
-//
-//trait ContainerBehavior {
-//  this : Actor =>
-//
-//  def ContainableObject : Container.ValidContainer
-//
-//  val containerBehavior : Receive = {
-//    case Container.GetMoveItem(where_src, destination, where_dest, other_item) =>
-//      val slot : EquipmentSlot = ContainableObject.Slot(where_src)
-//      val equipment = slot.Equipment
-//      slot.Equipment = None
-//      sender ! Container.GiveMoveItem(ContainableObject, where_src, equipment, destination, where_dest, other_item)
-//
-//    case Container.TakeMoveItem(source_index, destination, destination_index) =>
-//      val slot : EquipmentSlot = ContainableObject.Slot(source_index)
-//      val equipment : Option[Equipment] = slot.Equipment
-//      slot.Equipment = None
-//      sender ! (equipment match {
-//        case Some(item) =>
-//          Container.GivingMoveItem(item, ContainableObject, source_index, destination, destination_index)
-//        case None =>
-//          Container.NoMoveItem(ContainableObject, source_index)
-//      })
-//
-//    case Container.TakeMoveItem2(item_guid, destination, destination_index) =>
-//      ContainableObject.Find(item_guid) match {
-//        case Some(source_index) =>
-//          val slot : EquipmentSlot = ContainableObject.Slot(source_index)
-//          val equipment : Option[Equipment] = slot.Equipment
-//          slot.Equipment = None
-//          sender ! (equipment match {
-//            case Some(item) =>
-//              Container.GivingMoveItem(item, ContainableObject, source_index, destination, destination_index)
-//            case None => ;
-//          })
-//
-//        case None =>
-//          sender ! Container.NoMoveItem(ContainableObject, 65535)
-//      }
-//
-//    case Container.PutMoveItem(item, target_index, source, source_index) =>
-//      val slot : EquipmentSlot = ContainableObject.Slot(target_index)
-//      val equipment : Option[Equipment] = slot.Equipment
-//      if( {
-//        val tile = item.Definition.Tile
-//        ContainableObject.Collisions(target_index, tile.Width, tile.Height) match {
-//          case Success(Nil) => //no item swap
-//            true
-//          case Success(_ :: Nil) => //one item to swap
-//            true
-//          case Success(_) | Failure(_) =>
-//            false //abort when too many items at destination or other failure case
-//        }
-//      }) {
-//        slot.Equipment = None
-//        slot.Equipment = item
-//        equipment match {
-//          case Some(swapItem) =>
-//            sender ! Container.GivingMoveItem(swapItem, ContainableObject, target_index, source, source_index)
-//          case None => ;
-//        }
-//        sender ! Container.ItemMoved(item, ContainableObject, target_index)
-//      }
-//      else {
-//        sender ! Container.DropMoveItem(item, source, source_index)
-//      }
-//  }
-//}
